+++ Modules/Zoom/Database/Migrations/2020_10_26_084054_add_columns_zoom_meetings_table.php
@@ -13,6 +13,9 @@ return new class extends Migration
      */
     public function up()
     {
+        if (! Schema::hasTable('zoom_meetings')) {
+            return;
+        }
         Schema::table('zoom_meetings', function (Blueprint $table) {
             $table->unsignedBigInteger('source_meeting_id')->nullable();
             $table->foreign('source_meeting_id')->references('id')->on('zoom_meetings')->onDelete('cascade')->onUpdate('cascade');
